new_email_sender_v2.8/requirements.txt
tqdm>=4.65.0
dnspython>=2.3.0
tabulate>=0.9.0